refactor(scripts): reorganize build scripts
* `bin/<task>.js` -> `scripts/<task>.js`
* `lib/<module>` -> `scripts/_<module>/*`
* no change to logic other than corrections due to new file structure
