+++ scripts/cleanPublic.js
@@ -1,8 +1,11 @@
 #!/usr/bin/env node
 'use strict';
 
-const CONFIG = require('../_config');
+// TODO: migrate to scripts/clean.js
+
 const { remove } = require('fs-extra');
+
+const CONFIG = require('./_config');
 
 (async function () {
     await remove(CONFIG.publicDir);
