+++ scripts/publish.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 'use strict';
 
-const  { publishDocs } = require('../lib/publish');
+const { publishDocs } = require('./_publish/publishDocs');
 
 (async function () {
     await publishDocs();
